+++ src/app/Balloon.App.Api/v2/Nodes.php
@@ -23,6 +23,7 @@ use Balloon\Filesystem\Node\AttributeDecorator as NodeAttributeDecorator;
 use Balloon\Filesystem\Node\Collection;
 use Balloon\Filesystem\Node\File;
 use Balloon\Filesystem\Node\NodeInterface;
+use Balloon\Filesystem\Storage;
 use Balloon\Helper;
 use Balloon\Server;
 use Balloon\Server\User;
@@ -69,12 +70,20 @@ class Nodes extends Controller
     protected $node_decorator;
 
     /**
+     * Storage.
+     *
+     * @var Storage
+     */
+    protected $storage;
+
+    /**
      * Initialize.
      */
-    public function __construct(Server $server, NodeAttributeDecorator $decorator, LoggerInterface $logger)
+    public function __construct(Server $server, NodeAttributeDecorator $decorator, LoggerInterface $logger, Storage $storage)
     {
         $this->fs = $server->getFilesystem();
         $this->user = $server->getIdentity();
+        $this->storage = $storage;
         $this->server = $server;
         $this->node_decorator = $decorator;
         $this->logger = $logger;