+++ src/lib/Filesystem/Storage/Adapter/AdapterInterface.php
@@ -14,6 +14,8 @@ namespace Balloon\Filesystem\Storage\Adapter;
 use Balloon\Filesystem\Node\Collection;
 use Balloon\Filesystem\Node\File;
 use Balloon\Filesystem\Node\NodeInterface;
+use Balloon\Server\User;
+use MongoDB\BSON\ObjectId;
 
 interface AdapterInterface
 {
@@ -30,17 +32,19 @@ interface AdapterInterface
     /**
      * Get stored file.
      *
-     *
      * @return resource
      */
     public function getFile(File $file, array $attributes);
 
     /**
      * Store file.
-     *
-     * @param resource $contents
      */
-    public function storeFile(File $file, $contents);
+    public function storeFile(File $file, ObjectId $session);
+
+    /**
+     * Store temporary file.
+     */
+    public function storeTemporaryFile($stream, User $user, ?ObjectId $session): ObjectId;
 
     /**
      * Create collection.