+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2.0.3
+**Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
+**Date**: Tue Jul 10 10:02:39 CEST 2018
+
+* API: [CHANGE] quota attributes are now part of the root document GET /api/v2/users/{id}
+* CORE: [CHANGE] Implemented #92, Store chunks in gridfs temporary storage instead filesystem, increases stability and performance. Also fixes issues in load balanced environements.
+* CORE: [CHANGE] Quota is now checked after each uploaded chunk (implemented in #92)
+* CORE: [CHANGE] Added hook script to clean temporary blobs from gridfs #92
+* CORE: [CHANGE] Added mongodb index for group.member (Delta\CoreInstallation)
+
+
 ## 2.0.2
 **Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
 **Date**: Thu Jul 05 11:08:13 CEST 2018